requirements.txt
llama-index
faiss-cpu
llama-index-vector-stores-faiss
llama-index-embeddings-huggingface
llama-index-llms-ollama